move dir manager to utils
